refactor(editor): semantic media nodeview 분리
