+++ front/e2e/editor-studio-state.spec.ts
@@ -367,6 +367,41 @@ test.describe("editor studio state", () => {
     const mermaidNodeViewSource = existsSync(mermaidNodeViewPath)
       ? readFileSync(mermaidNodeViewPath, "utf8")
       : ""
+    const editorNodeViewSharedPath = path.resolve(__dirname, "../src/components/editor/editorNodeViewShared.ts")
+    expect(existsSync(editorNodeViewSharedPath)).toBe(true)
+    const editorNodeViewSharedSource = existsSync(editorNodeViewSharedPath)
+      ? readFileSync(editorNodeViewSharedPath, "utf8")
+      : ""
+    const calloutNodeViewPath = path.resolve(__dirname, "../src/components/editor/calloutNodeView.tsx")
+    expect(existsSync(calloutNodeViewPath)).toBe(true)
+    const calloutNodeViewSource = existsSync(calloutNodeViewPath)
+      ? readFileSync(calloutNodeViewPath, "utf8")
+      : ""
+    const toggleNodeViewPath = path.resolve(__dirname, "../src/components/editor/toggleNodeView.tsx")
+    expect(existsSync(toggleNodeViewPath)).toBe(true)
+    const toggleNodeViewSource = existsSync(toggleNodeViewPath)
+      ? readFileSync(toggleNodeViewPath, "utf8")
+      : ""
+    const linkCardNodeViewsPath = path.resolve(__dirname, "../src/components/editor/linkCardNodeViews.tsx")
+    expect(existsSync(linkCardNodeViewsPath)).toBe(true)
+    const linkCardNodeViewsSource = existsSync(linkCardNodeViewsPath)
+      ? readFileSync(linkCardNodeViewsPath, "utf8")
+      : ""
+    const formulaNodeViewsPath = path.resolve(__dirname, "../src/components/editor/formulaNodeViews.tsx")
+    expect(existsSync(formulaNodeViewsPath)).toBe(true)
+    const formulaNodeViewsSource = existsSync(formulaNodeViewsPath)
+      ? readFileSync(formulaNodeViewsPath, "utf8")
+      : ""
+    const rawMarkdownNodeViewPath = path.resolve(__dirname, "../src/components/editor/rawMarkdownNodeView.tsx")
+    expect(existsSync(rawMarkdownNodeViewPath)).toBe(true)
+    const rawMarkdownNodeViewSource = existsSync(rawMarkdownNodeViewPath)
+      ? readFileSync(rawMarkdownNodeViewPath, "utf8")
+      : ""
+    const resizableImageNodeViewPath = path.resolve(__dirname, "../src/components/editor/resizableImageNodeView.tsx")
+    expect(existsSync(resizableImageNodeViewPath)).toBe(true)
+    const resizableImageNodeViewSource = existsSync(resizableImageNodeViewPath)
+      ? readFileSync(resizableImageNodeViewPath, "utf8")
+      : ""
     const blockSelectionModelSource = readFileSync(
       path.resolve(__dirname, "../src/components/editor/blockSelectionModel.ts"),
       "utf8"
@@ -908,6 +943,40 @@ test.describe("editor studio state", () => {
     expect(codeBlockNodeViewSource).toContain("export const CodeBlockEditorStyles =")
     expect(mermaidNodeViewSource).toContain("export const MermaidBlockView =")
     expect(mermaidNodeViewSource).toContain("export const MermaidEditorStyles =")
+    expect(editorExtensionsSource).toContain('from "./calloutNodeView"')
+    expect(editorExtensionsSource).toContain('from "./toggleNodeView"')
+    expect(editorExtensionsSource).toContain('from "./linkCardNodeViews"')
+    expect(editorExtensionsSource).toContain('from "./formulaNodeViews"')
+    expect(editorExtensionsSource).toContain('from "./rawMarkdownNodeView"')
+    expect(editorExtensionsSource).toContain('from "./resizableImageNodeView"')
+    expect(editorExtensionsSource.split("\n").length).toBeLessThanOrEqual(900)
+    expect(editorExtensionsSource).not.toContain("const CalloutBlockView =")
+    expect(editorExtensionsSource).not.toContain("const ToggleBlockView =")
+    expect(editorExtensionsSource).not.toContain("const LinkCardEditorView =")
+    expect(editorExtensionsSource).not.toContain("const FormulaBlockView =")
+    expect(editorExtensionsSource).not.toContain("const InlineFormulaView =")
+    expect(editorExtensionsSource).not.toContain("const RawMarkdownBlockView =")
+    expect(editorExtensionsSource).not.toContain("const ResizableImageView =")
+    expect(editorExtensionsSource).not.toContain("const CalloutEditorWrapper = styled")
+    expect(editorExtensionsSource).not.toContain("const ImageBlockWrapper = styled")
+    expect(editorNodeViewSharedSource).toContain("export const selectDomTextContents =")
+    expect(editorNodeViewSharedSource).toContain("export const useAutosizeTextarea =")
+    expect(editorNodeViewSharedSource).toContain("export const useDebouncedAttributeCommit =")
+    expect(calloutNodeViewSource).toContain("export const CalloutBlockView =")
+    expect(calloutNodeViewSource).toContain("export const CalloutNodeViewStyles =")
+    expect(toggleNodeViewSource).toContain("export const ToggleBlockView =")
+    expect(toggleNodeViewSource).toContain("export const ToggleNodeViewStyles =")
+    expect(linkCardNodeViewsSource).toContain("export const BookmarkBlock =")
+    expect(linkCardNodeViewsSource).toContain("export const EmbedBlock =")
+    expect(linkCardNodeViewsSource).toContain("export const FileBlock =")
+    expect(linkCardNodeViewsSource).toContain("export const LinkCardNodeViewStyles =")
+    expect(formulaNodeViewsSource).toContain("export const FormulaBlockView =")
+    expect(formulaNodeViewsSource).toContain("export const InlineFormulaView =")
+    expect(formulaNodeViewsSource).toContain("export const FormulaNodeViewStyles =")
+    expect(rawMarkdownNodeViewSource).toContain("export const RawMarkdownBlockView =")
+    expect(rawMarkdownNodeViewSource).toContain("export const RawMarkdownNodeViewStyles =")
+    expect(resizableImageNodeViewSource).toContain("export const ResizableImageView =")
+    expect(resizableImageNodeViewSource).toContain("export const ResizableImageNodeViewStyles =")
     expect(blockEditorEngineLayersSource).toContain("export const BlockEditorToolbarLayer =")
     expect(blockEditorEngineLayersSource).toContain("export const BlockEditorQuickInsertLayer =")
     expect(blockEditorEngineLayersSource).toContain("export const BlockEditorSlashMenuLayer =")
